Implement #700.
